refactor(remote_transfer): split tests to separate file (#1266)
Splits src/remote_transfer.rs (401 LOC) into:
- src/remote_transfer/mod.rs   (~205 LOC) — production code
- src/remote_transfer/tests.rs (~195 LOC) — test module body

Pure refactor: zero logic changes, all 6 tests pass unmodified.
Brings the module below the 400-line rule.

Closes part of #1266 (1/70 modules).
