+++ .github/hooks/amplihack-hooks.json
@@ -1,47 +1,47 @@
 {
-  "version": 1,
   "hooks": {
-    "sessionStart": [
+    "agentStop": [
       {
-        "type": "command",
-        "bash": ".github/hooks/session-start",
-        "timeoutSec": 30
-      }
-    ],
-    "sessionEnd": [
-      {
-        "type": "command",
-        "bash": ".github/hooks/session-stop",
-        "timeoutSec": 30
-      }
-    ],
-    "userPromptSubmitted": [
-      {
-        "type": "command",
-        "bash": ".github/hooks/user-prompt-submit",
-        "timeoutSec": 10
-      }
-    ],
-    "preToolUse": [
-      {
-        "type": "command",
-        "bash": ".github/hooks/pre-tool-use",
-        "timeoutSec": 15
+        "bash": "/home/azureuser/src/Simard/.github/hooks/stop",
+        "timeoutSec": 30,
+        "type": "command"
       }
     ],
     "postToolUse": [
       {
-        "type": "command",
-        "bash": ".github/hooks/post-tool-use",
-        "timeoutSec": 10
+        "bash": "/home/azureuser/src/Simard/.github/hooks/post-tool-use",
+        "timeoutSec": 30,
+        "type": "command"
       }
     ],
-    "errorOccurred": [
+    "preCompact": [
       {
-        "type": "command",
-        "bash": ".github/hooks/error-occurred",
-        "timeoutSec": 10
+        "bash": "/home/azureuser/src/Simard/.github/hooks/pre-compact",
+        "timeoutSec": 30,
+        "type": "command"
+      }
+    ],
+    "preToolUse": [
+      {
+        "bash": "/home/azureuser/src/Simard/.github/hooks/pre-tool-use",
+        "timeoutSec": 30,
+        "type": "command"
+      }
+    ],
+    "sessionStart": [
+      {
+        "bash": "/home/azureuser/src/Simard/.github/hooks/session-start",
+        "timeoutSec": 30,
+        "type": "command"
+      }
+    ],
+    "userPromptSubmitted": [
+      {
+        "bash": "/home/azureuser/src/Simard/.github/hooks/user-prompt-submit",
+        "timeoutSec": 30,
+        "type": "command"
       }
     ]
-  }
+  },
+  "version": 1
 }
