+++ .github/hooks/pre-tool-use
@@ -1,67 +1,20 @@
 #!/usr/bin/env bash
-# Copilot hook wrapper - generated by amplihack (rust engine)
-# Aggregates amplihack and XPIA pre-tool validation into one JSON response
-REPO_ROOT="$(git rev-parse --show-toplevel 2>/dev/null)" || REPO_ROOT="$(pwd)"
-INPUT=$(cat)
+set -euo pipefail
 
-AMPLIHACK_OUTPUT=$(echo "$INPUT" | /home/azureuser/.local/bin/amplihack-hooks pre-tool-use "$@" 2>/dev/null || printf '{}')
+HOOKS_BIN=""
+for candidate in \
+    "$(command -v amplihack-hooks 2>/dev/null)" \
+    "${HOME}/.amplihack/target/release/amplihack-hooks" \
+    "${HOME}/.amplihack/target/debug/amplihack-hooks"; do
+  if [ -n "$candidate" ] && [ -x "$candidate" ]; then
+    HOOKS_BIN="$candidate"
+    break
+  fi
+done
 
-XPIA_OUTPUT="{}"
-XPIA_HOOKS="$HOME/.amplihack/.claude/tools/xpia/hooks"
-if [[ -f "${XPIA_HOOKS}/pre_tool_use.py" ]]; then
-    XPIA_OUTPUT=$(echo "$INPUT" | python3 "${XPIA_HOOKS}/pre_tool_use.py" "$@" 2>/dev/null || printf '{}')
-elif [[ -n "$REPO_ROOT" ]] && [[ -f "${REPO_ROOT}/.claude/tools/xpia/hooks/pre_tool_use.py" ]]; then
-    XPIA_OUTPUT=$(echo "$INPUT" | python3 "${REPO_ROOT}/.claude/tools/xpia/hooks/pre_tool_use.py" "$@" 2>/dev/null || printf '{}')
+if [ -z "$HOOKS_BIN" ]; then
+  echo "[amplihack] hooks binary not found, skipping" >&2
+  exit 0
 fi
 
-python3 - "$AMPLIHACK_OUTPUT" "$XPIA_OUTPUT" <<'PY'
-import json
-import sys
-
-
-def parse_payload(raw: str) -> dict:
-    raw = raw.strip()
-    if not raw:
-        return {}
-    for line in reversed(raw.splitlines()):
-        line = line.strip()
-        if not line:
-            continue
-        try:
-            value = json.loads(line)
-        except json.JSONDecodeError:
-            continue
-        if isinstance(value, dict):
-            return value
-    return {}
-
-
-amplihack = parse_payload(sys.argv[1])
-xpia = parse_payload(sys.argv[2])
-
-permission = xpia.get("permissionDecision")
-if permission in {"allow", "deny", "ask"}:
-    print(json.dumps(xpia))
-    raise SystemExit(0)
-
-permission = amplihack.get("permissionDecision")
-if permission in {"allow", "deny", "ask"}:
-    print(json.dumps(amplihack))
-    raise SystemExit(0)
-
-if amplihack.get("block"):
-    print(
-        json.dumps(
-            {
-                "permissionDecision": "deny",
-                "message": amplihack.get(
-                    "message",
-                    "Blocked by amplihack pre-tool-use hook.",
-                ),
-            }
-        )
-    )
-    raise SystemExit(0)
-
-print("{}")
-PY
+exec "$HOOKS_BIN" pre-tool-use "$@"
