+++ .github/hooks/user-prompt-submit
@@ -1,14 +1,21 @@
 #!/usr/bin/env bash
-# Copilot hook wrapper - generated by amplihack (rust engine)
-# Runs multiple hook scripts for this event
-INPUT=$(cat)
+set -euo pipefail
 
-echo "$INPUT" | /home/azureuser/.local/bin/amplihack-hooks user-prompt-submit "$@" || true
+HOOKS_BIN=""
+for candidate in \
+    "$(command -v amplihack-hooks 2>/dev/null)" \
+    "${HOME}/.amplihack/target/release/amplihack-hooks" \
+    "${HOME}/.amplihack/target/debug/amplihack-hooks"; do
+  if [ -n "$candidate" ] && [ -x "$candidate" ]; then
+    HOOKS_BIN="$candidate"
+    break
+  fi
+done
 
-AMPLIHACK_HOOKS="$HOME/.amplihack/.claude/tools/amplihack/hooks"
-REPO_ROOT="$(git rev-parse --show-toplevel 2>/dev/null)" || REPO_ROOT="$(pwd)"
-if [[ -f "${AMPLIHACK_HOOKS}/workflow_classification_reminder.py" ]]; then
-    echo "$INPUT" | python3 "${AMPLIHACK_HOOKS}/workflow_classification_reminder.py" "$@" 2>/dev/null || true
-elif [[ -n "$REPO_ROOT" ]] && [[ -f "${REPO_ROOT}/.claude/tools/amplihack/hooks/workflow_classification_reminder.py" ]]; then
-    echo "$INPUT" | python3 "${REPO_ROOT}/.claude/tools/amplihack/hooks/workflow_classification_reminder.py" "$@" 2>/dev/null || true
+if [ -z "$HOOKS_BIN" ]; then
+  echo "[amplihack] hooks binary not found, skipping" >&2
+  exit 0
 fi
+
+"$HOOKS_BIN" workflow-classification-reminder "$@" || true
+"$HOOKS_BIN" user-prompt-submit "$@" || true
